refactor(sight): address token savings review feedback
- Fix savings rate formula inconsistency: unify get_session_savings
  to use compounded/total_tokens (same as get_token_savings list page)
- Fix duplicate pie chart slices: aggregate strategy_breakdown by
  label instead of raw operation, so unknown ops merge into one slice
- Fix full table scan: add get_session_for_savings() method that
  queries by session_id directly instead of list_all + .find
- Add classification comments explaining compress-toon/compress-schema
  bucket assignment rationale
- Extract token_savings.rs module from handlers.rs (~800 lines) to
  reduce file size and improve maintainability
